move drawer header into settings and notif components, use type instead of type name to check if drawer is open
