+++ apps/web/src/contexts/globalLayout/GlobalSidebar/SidebarDrawerContext.tsx
@@ -48,35 +48,18 @@ export const useDrawerActions = (): DrawerActions => {
 
 type Props = { children: ReactNode };
 
-function isDrawerType(type: string): type is DrawerType {
-  console.log({ type });
-  return type === 'Notifications' || type === 'Settings';
-}
-
 function SidebarDrawerProvider({ children }: Props): ReactElement {
   const [drawerState, setDrawerState] = useState<DrawerState>({ activeDrawer: null });
 
   const showDrawer = useCallback((props: ActiveDrawer) => {
     setDrawerState((previous) => {
-      // infer the drawer type from the content, and check that it's a valid type
-      console.log(
-        props.content.type,
-        props.content.type.name,
-        isDrawerType(props.content.type.name)
-      );
-      const propsDrawerType = typeof props.content.type !== 'string' ? props.content.type.name : '';
-
-      if (!isDrawerType(propsDrawerType)) {
-        throw new Error('Invalid drawer type!');
-      }
-
       // if the same drawer type is already open, close it
-      if (previous.activeDrawer?.drawerType === propsDrawerType) {
+      if (previous.activeDrawer?.content.type === props.content.type) {
         return { activeDrawer: null };
       }
 
-      // otherwise, open the new drawer. save the inferred drawer type
-      return { activeDrawer: { ...props, drawerType: propsDrawerType } };
+      // otherwise, open the new drawer
+      return { activeDrawer: props };
     });
   }, []);
